Update Library_Management_System.py
the Library class is a central controller for a library management system that handles adding, searching, borrowing and returning books using efficient data structures.

it stores books in a hash table for fast lookups by ISBN and a (BST) for sorting and searching by title. When a new book is added, it’s inserted into both structures.

the class allows users to check out books, manages waitlists through a queue for books already borrowed, and automatically assigns the book to the next user when returned.

it also provides search functions to find books by ISBN or title and lists all books in alphabetical order.